extracted BrowseMenu class
