+++ src/gui/gui.py
@@ -5,6 +5,7 @@ from tkinter.ttk import Label
 
 from core.directory_processors import FlatDirectoryProcessor
 from resources.watermarks import DEFAULT_DARK_WATERMARK, DEFAULT_LIGHT_WATERMARK
+from .browse_menu import BrowseMenu
 from .config_menu import ConfigMenu
 
 
@@ -29,23 +30,11 @@ class MainBody(ttk.Frame):
         self.main_label = Label(self, text="Add watermarks", font=('Helvetica', 30))
         self.main_label.grid(row=0, column=0)
 
-        self.browse_btn = ttk.Button(self, text='Browse', command=self.browse_directories)
-        self.browse_btn.grid(row=1, column=0, sticky=tk.W)
-
-        self.directory = tk.StringVar()
-        self.directory.set('directory containing photos')
-        self.directory_info = ttk.Entry(self, textvariable=self.directory)
-        self.directory_info.grid(row=1, column=1, columnspan=3, sticky=tk.W)
-
-        self.add_btn = ttk.Button(self, text='ADD', command=lambda: self.watermark_photos())
-        self.add_btn.grid(row=1, column=3)
+        self.browse_menu = BrowseMenu(self, self.watermark_photos)
+        self.browse_menu.grid(row=2, columnspan=2)
 
         self.config_menu = ConfigMenu(self)
-        self.config_menu.grid(row=3, column=0)
-
-    def browse_directories(self):
-        photos_directory = filedialog.askdirectory()
-        self.directory.set(photos_directory)
+        self.config_menu.grid(row=3, column=0, sticky=tk.W, padx=25)
 
     def watermark_photos(self):
         processor = FlatDirectoryProcessor(
@@ -57,4 +46,4 @@ class MainBody(ttk.Frame):
             corners=self.config_menu.get_corners()
         )
 
-        processor.handle_directory(self.directory.get())
+        processor.handle_directory(self.browse_menu.get_directory())
